Create package folder and move helper functions
